requirements.txt
llama-index==0.10.20
openai==1.2.4
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
